requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
SQLAlchemy==2.0.43
WTForms==3.2.1
email-validator==2.3.0
python-dotenv==1.2.2
psycopg2-binary==2.9.12
pg8000==1.31.2
Werkzeug==3.1.3
